SWDEV-440004 - Fix Unit_hipHostMalloc_AllocateUseMoreThanAvailGPUMemory
int(32bit -2147483648～2147483647) could overflow if allocated size >= 2GB(2147483648).
Use size_t instead.



[ROCm/hip-tests commit: c7aa062b66dab628e6f4813b938638413e5b1120]

+++ catch/unit/memory/hipHostMalloc.cc
@@ -48,7 +48,7 @@ static constexpr size_t sizeBytes{numElements * sizeof(int)};
 
 #if HT_AMD
 static __global__ void kerTestMemAccess(char *buf) {
-  int myId = threadIdx.x + blockDim.x * blockIdx.x;
+  size_t myId = threadIdx.x + blockDim.x * blockIdx.x;
   buf[myId] = VALUE;
 }
 #endif
